workaround for watchfd

+++ gui.h
@@ -20,8 +20,8 @@ public:
 	FSChar *selectDirectory(const char *question, const FSChar *start);
 	bool askYesNo(const char *choice);
 	void error(const char *error);
-	void watchfd(int fd, void callback(int fd, void *callback_data), void *callback_data);
-	void unwatchfd(int fd);
+	void watchfd(long int fd, void callback(long int fd, void *callback_data), void *callback_data);
+	void unwatchfd(long int fd);
 private:
 	class GUIPrivate;
 	GUIPrivate *priv;
